backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
praw>=7.7.1
pydantic>=2.4.2
python-multipart>=0.0.6
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4